add bank user service module and modify unreasonable designs for User, credit card and bank user service facade.

+++ dataModel/src/main/java/com/stori/datamodel/repository/CreditCardRepository.java
@@ -3,5 +3,5 @@ package com.stori.datamodel.repository;
 import com.stori.datamodel.model.CreditCard;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CreditCardRepository extends CrudRepository<CreditCard, Integer> {
+public interface CreditCardRepository extends CrudRepository<CreditCard, Long> {
 }
