+++ dataModel/src/main/java/com/stori/datamodel/repository/MerchantRepository.java
@@ -3,5 +3,5 @@ package com.stori.datamodel.repository;
 import com.stori.datamodel.model.Merchant;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MerchantRepository extends CrudRepository<Merchant, Integer> {
+public interface MerchantRepository extends CrudRepository<Merchant, Long> {
 }
