+++ dataModel/src/main/java/com/stori/datamodel/repository/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.stori.datamodel.repository;
 import com.stori.datamodel.model.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<Transaction, Integer> {
+public interface TransactionRepository extends CrudRepository<Transaction, Long> {
 }
